+++ AllMyLights/Connectors/Sinks/ISink.cs
@@ -5,6 +5,6 @@ namespace AllMyLights.Connectors.Sinks
     public interface ISink
     {
         void Consume(object value);
-        IEnumerable<object> GetConsumers() => new List<object>();
+        object GetInfo() => null;
     }
 }